evaluacion/requirements.txt
flask==2.0.1
mysql-connector-python
